+++ lib/bootstrap_form/form_builder.rb
@@ -246,7 +246,8 @@ module BootstrapForm
       name = args.first
 
       options[:class] = ["form-group", options[:class]].compact.join(' ')
-      options[:class] << " row" if get_group_layout(options[:layout]) == :horizontal
+      options[:class] << " row" if get_group_layout(options[:layout]) == :horizontal &&
+                                   !options[:class].split.include?("form-row")
       options[:class] << " form-inline" if field_inline_override?(options[:layout])
       options[:class] << " #{feedback_class}" if options[:icon]